Add flush(), associated unit tests, resolve() checks if a bind exists
It was expected that resolve() would return an already known instance if it could, however, it was not taking binds into account. As such, it will check if a bind exists before an already resolved instance and return accordingly.

With that in mind, it may be a good idea to change resolve() to not necessarily always assume to cache the  resolved item to be returned on future calls. Instead it might be a good idea to create a new bag for singletons & instances.

For now, satisfy the current expectations.