Can now pass parameters to make()/resolve()
User-defined parameters can be passed into make()/resolve() to better control setting specific parameters to the object that is receiving injected dependencies.

For now, the parameters array will not be forwarded to dependencies at all. Obviously this is something that should be considered for the future, but needs more time to plan out a decent approach.